% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmwpowd.R
\name{wmwpowd}
\alias{wmwpowd}
\title{Power Calculation with wmwpowd}
\usage{
wmwpowd(n, m, distn, distm, sides, alpha = 0.05, nsims = 10000)
}
\arguments{
\item{n}{Sample size for the first distribution (numeric)}

\item{m}{Sample size for the second distribution (numeric)}

\item{alpha}{Type I error rate or significance level (numeric)}

\item{distn}{Base R’s name for the first distribution and any required parameters 
("norm", "beta", "cauchy", "f", "gamma", "lnorm", "unif", "weibull","exp", "chisq", "t", "doublex")}

\item{distm}{Base R’s name for the second distribution and any required parameters 
("norm", "beta", "cauchy", "f", "gamma", "lnorm", "unif", "weibull","exp", "chisq", "t", "doublex")}

\item{sides}{Options are “two.sided”, “less”, or “greater”. “less” means the
alternative hypothesis is that distn is less than distm (string)}

\item{nsims}{Number of simulated datasets for determining power; 10,000 is the default (numeric)}
}
\description{
\emph{wmwpowd} has two purposes:

1. Determine the power for a one-sided or two-sided Wilcoxon-Mann-Whitney test
with an exact p-value given two user specified distributions.

2. Determine p, the P(X<Y), where X represents random draws from one continuous
probability distribution and Y represents random draws from another distribution;
p is useful for quantifying the effect size that the Wilcoxon-Mann-Whitney test is
assessing.

Both 1. and 2. are calculated empirically using simulated data and output automatically.
}
\note{
Example of distn, distm: “norm(1,2)” or “exp(1)”

In addition to all continuous distributions supported in Base R, \emph{wmwpowd} also supports the 
double exponential distribution from the smoothmest package

The output WMWOdds is p expressed as odds (p/(1-p))

Use $ notation to select specific output parameters

The function has been optimized to run through simulations quickly; long wait times are unlikely 
for n and m of 50 or fewer
}
\examples{
# 1. We want to determine the statistical power to compare body length measured on two groups of
# rabbits. Each group (X and Y) has 7 rabbits. We assume that body length will be normally 
# distributed and have a constant standard deviation of 2 cm among groups. We assume that Group X 
# will have a mean of 35 cm and Group Y will have a mean of 32 cm; the desired type I error is 0.05.

\donttest{wmwpowd(n = 7, m = 7, distn = "norm(35,2)", distm = "norm(32,2)", sides = "two.sided", 
        alpha = 0.05, nsims=10000)}
        
\dontshow{wmwpowd(n = 5, m = 5, distn = "norm(35,2)", distm = "norm(32,2)", sides = "two.sided", 
        alpha = 0.05, nsims=100)}

# 2. We are interested in determining the statistical power (with type I error = 0.05) for a
# comparison of the use of ornamentation among fiddle players living in two regions of the United 
# States: X county, Texas and Y county, North Carolina. A random sample of 18 fiddlers will be 
# collected within each state. The fiddlers will practice and perform a standardized version of the 
# Tennessee Waltz. The proportion of melody notes that are ornamented (including vibrato) will be 
# calculated. We assume that the proportion will follow a beta distribution with a mean of 0.40 and 
# a shape well described by alpha = 8 and beta = 12 among the Texas fiddlers. We assume that 
# the distribution will be shifted to a lower mean of 0.25 and have the shape alpha = 2, 
# beta = 6 for the North Carolina fiddlers.

\donttest{wmwpowd(n=18, m=18, distn = "beta(8,12)", distm = "beta(2,6)", sides = "two.sided", 
        alpha = 0.05, nsims=10000)}
        
\dontshow{wmwpowd(n=5, m=5, distn = "beta(8,12)", distm = "beta(2,6)", sides = "two.sided", 
        alpha = 0.05, nsims=100)}
}
